+++ bindings/c/src/aircraft/cg_limit.rs
@@ -19,21 +19,11 @@ use efb::measurements::{Length, Mass};
 /// Returns the limit's mass.
 #[no_mangle]
 pub extern "C" fn efb_cg_limit_mass(limit: &CGLimit) -> &Mass {
-    &limit.mass
-}
-
-#[no_mangle]
-pub extern "C" fn efb_cg_limit_set_mass(limit: &mut CGLimit, mass: Mass) {
-    limit.mass = mass
+    limit.mass()
 }
 
 /// Returns the limit's distance in reference to the aircraft's datum.
 #[no_mangle]
 pub extern "C" fn efb_cg_limit_distance(limit: &CGLimit) -> &Length {
-    &limit.distance
-}
-
-#[no_mangle]
-pub extern "C" fn efb_cg_limit_set_distance(limit: &mut CGLimit, distance: Length) {
-    limit.distance = distance
+    limit.distance()
 }