+++ efb/tests/aircraft_mb_test.rs
@@ -15,21 +15,15 @@
 
 use efb::aircraft::{Aircraft, CGEnvelope, CGLimit, FuelTank, Station};
 use efb::measurements::{Length, Mass, Volume};
-use efb::{diesel, Fuel, FuelType};
+use efb::{Fuel, FuelType, diesel};
 
 /// Returns the an aircraft we use for the tests.
 fn aircraft() -> Aircraft {
     Aircraft {
         registration: String::from("N12345"),
         stations: vec![
-            Station {
-                arm: Length::m(1.0),
-                description: None,
-            },
-            Station {
-                arm: Length::m(2.0),
-                description: None,
-            },
+            Station::new(Length::m(1.0), None),
+            Station::new(Length::m(2.0), None),
         ],
         empty_mass: Mass::kg(800.0),
         empty_balance: Length::m(1.0),
@@ -38,40 +32,16 @@ fn aircraft() -> Aircraft {
             // To spice things up, lets test with two tanks in the wings
             // configured as separate tanks and an additional tank in the
             // aft cargo compartment.
-            FuelTank {
-                capacity: Volume::l(50.0),
-                arm: Length::m(1.0),
-            },
-            FuelTank {
-                capacity: Volume::l(50.0),
-                arm: Length::m(1.0),
-            },
-            FuelTank {
-                capacity: Volume::l(20.0),
-                arm: Length::m(1.5),
-            },
+            FuelTank::new(Volume::l(50.0), Length::m(1.0)),
+            FuelTank::new(Volume::l(50.0), Length::m(1.0)),
+            FuelTank::new(Volume::l(20.0), Length::m(1.5)),
         ],
         cg_envelope: CGEnvelope::new(vec![
-            CGLimit {
-                mass: Mass::kg(0.0),
-                distance: Length::m(1.0),
-            },
-            CGLimit {
-                mass: Mass::kg(800.0),
-                distance: Length::m(1.0),
-            },
-            CGLimit {
-                mass: Mass::kg(1000.0),
-                distance: Length::m(1.0),
-            },
-            CGLimit {
-                mass: Mass::kg(1000.0),
-                distance: Length::m(1.5),
-            },
-            CGLimit {
-                mass: Mass::kg(0.0),
-                distance: Length::m(1.5),
-            },
+            CGLimit::new(Mass::kg(0.0), Length::m(1.0)),
+            CGLimit::new(Mass::kg(800.0), Length::m(1.0)),
+            CGLimit::new(Mass::kg(1000.0), Length::m(1.0)),
+            CGLimit::new(Mass::kg(1000.0), Length::m(1.5)),
+            CGLimit::new(Mass::kg(0.0), Length::m(1.5)),
         ]),
         notes: None,
     }